Revert "feat: Fixed writeFile bug, refines Bmclapi mirror, de-async stage 10/17"
This reverts commit b78878817a07221fae1fbce9986942da646202ef.

+++ src/backend_migrate/util/File.ts
@@ -62,8 +62,12 @@ export function readFileAsJson(filePath: string): JSON{
 }
 
 export function writeFile(filePath: string, content: string): boolean{
-    fs.writeFileSync(filePath, content);
-    return exist(filePath);
+    if(exist(filePath)){
+        fs.writeFileSync(filePath, content);
+        return true;
+    }else{
+        return false;
+    }
 }
 
 export function removeFile(path: string): boolean{
